Segregating source code from header file: ContourRectangles.
