+++ src/.vscode/tasks.json
@@ -10,6 +10,7 @@
                 "-std=c++17",
                 "./Rectangle_Detection/main.cpp",
                 "./Detect_Rectangles/ContourRectangles.hpp",
+                "./Detect_Rectangles/ContourRectangles.cpp",
                 "./Detect_Rectangles/GaussianBlurTrackbar.hpp",
                 "./Detect_Rectangles/CannyThresholdTrackbar.hpp",
                 "./Detect_Rectangles/GaussianBlurTrackbar.cpp",
